Refactor Client generators to use Base library
