+++ eng/AnalyzerConfigOptionsExtensions.cs
