+++ generators/dotnetstandard-client/PrincipleStudios.OpenApiCodegen.Client.Analyzers/Analyzer.targets
@@ -28,10 +28,10 @@
 		<Compile Include="$(RepositoryEngineeringDir)MaybeNullWhenAttribute.cs" Visible="false" />
 		<Compile Include="$(RepositoryEngineeringDir)NotNullWhenAttribute.cs" Visible="false" />
 		<Compile Include="$(RepositoryEngineeringDir)BaseGenerator.cs" Link="Shared/BaseGenerator.cs" />
+		<Compile Include="$(RepositoryEngineeringDir)AnalyzerConfigOptionsExtensions.cs" Link="Shared/AnalyzerConfigOptionsExtensions.cs" />
 	</ItemGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="$(SolutionRoot)lib\PrincipleStudios.OpenApi.CSharp\PrincipleStudios.OpenApi.CSharp.csproj" SpecificVersion="true" />
-		<ProjectReference Include="$(SolutionRoot)lib\PrincipleStudios.OpenApiCodegen\PrincipleStudios.OpenApiCodegen.csproj" SpecificVersion="true" />
+		<ProjectReference Include="$(SolutionRoot)generators\dotnetstandard-client\PrincipleStudios.OpenApiCodegen.Client.Base\PrincipleStudios.OpenApiCodegen.Client.Base.csproj" SpecificVersion="true" />
 	</ItemGroup>
 </Project>
