--- generators/dotnetstandard-client/PrincipleStudios.OpenApiCodegen.Client.Analyzers/AnalyzerConfigOptionsExtensions.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Microsoft.CodeAnalysis.Diagnostics
-{
-    public static class AnalyzerConfigOptionsExtensions
-    {
-        public static string? GetAdditionalFilesMetadata(this AnalyzerConfigOptions opt, string fieldName) =>
-            opt.TryGetValue($"build_metadata.additionalfiles.{fieldName}", out var value) ? value : null;
-    }
-}
