+++ generators/dotnetstandard-client/PrincipleStudios.OpenApiCodegen.Client.Base/CSharpClientTransformer.cs
