+++ generators/dotnetstandard-client/PrincipleStudios.OpenApiCodegen.Client.Base/ControllerHandlebarsTemplateProcess.cs
