+++ generators/dotnetstandard-client/PrincipleStudios.OpenApiCodegen.Client.Base/OperationVisitor.cs
