+++ generators/dotnetstandard-client/PrincipleStudios.OpenApiCodegen.Client.Base/PrincipleStudios.OpenApiCodegen.Client.Base.csproj
@@ -1,0 +1,21 @@
+<Project Sdk="Microsoft.NET.Sdk">
+	<PropertyGroup>
+		<RootNamespace>PrincipleStudios.OpenApi.CSharp</RootNamespace>
+		<AssemblyName>PrincipleStudios.OpenApiCodegen.Client.Base</AssemblyName>
+		<TargetFramework>netstandard2.0</TargetFramework>
+		<VersionPrefix>$(SharedAnalyzerLibrariesVersion)</VersionPrefix>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<None Remove="templates\*.handlebars" />
+		<EmbeddedResource Include="templates\*.handlebars" />
+		<Compile Include="$(RepositoryEngineeringDir)IsExternalInit.cs" Visible="false" />
+		<Compile Include="$(RepositoryEngineeringDir)MaybeNullWhenAttribute.cs" Visible="false" />
+		<Compile Include="$(RepositoryEngineeringDir)NotNullWhenAttribute.cs" Visible="false" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="$(SolutionRoot)lib\PrincipleStudios.OpenApi.CSharp\PrincipleStudios.OpenApi.CSharp.csproj" SpecificVersion="true" />
+		<ProjectReference Include="$(SolutionRoot)lib\PrincipleStudios.OpenApiCodegen\PrincipleStudios.OpenApiCodegen.csproj" SpecificVersion="true" />
+	</ItemGroup>
+</Project>
