+++ generators/dotnetstandard-client/PrincipleStudios.OpenApiCodegen.Client.Base/templates/AddServices.cs
