+++ generators/dotnetstandard-client/PrincipleStudios.OpenApiCodegen.Client.Base/templates/ClientTemplate.cs
