+++ generators/dotnetstandard-client/PrincipleStudios.OpenApiCodegen.Client.Base/templates/addServices.handlebars
