+++ generators/dotnetstandard-client/PrincipleStudios.OpenApiCodegen.Client.Base/templates/client.handlebars
