+++ openapi-codegen.sln
@@ -57,6 +57,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PrincipleStudios.OpenApiCod
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PrincipleStudios.OpenApiCodegen", "lib\PrincipleStudios.OpenApiCodegen\PrincipleStudios.OpenApiCodegen.csproj", "{D1702287-F937-4429-BF77-4A3862746C1F}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PrincipleStudios.OpenApiCodegen.Client.Base", "generators\dotnetstandard-client\PrincipleStudios.OpenApiCodegen.Client.Base\PrincipleStudios.OpenApiCodegen.Client.Base.csproj", "{11554C34-D81C-4F07-A38D-1E1995014AD2}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -307,6 +309,18 @@ Global
 		{D1702287-F937-4429-BF77-4A3862746C1F}.Release|x64.Build.0 = Release|Any CPU
 		{D1702287-F937-4429-BF77-4A3862746C1F}.Release|x86.ActiveCfg = Release|Any CPU
 		{D1702287-F937-4429-BF77-4A3862746C1F}.Release|x86.Build.0 = Release|Any CPU
+		{11554C34-D81C-4F07-A38D-1E1995014AD2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{11554C34-D81C-4F07-A38D-1E1995014AD2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{11554C34-D81C-4F07-A38D-1E1995014AD2}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{11554C34-D81C-4F07-A38D-1E1995014AD2}.Debug|x64.Build.0 = Debug|Any CPU
+		{11554C34-D81C-4F07-A38D-1E1995014AD2}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{11554C34-D81C-4F07-A38D-1E1995014AD2}.Debug|x86.Build.0 = Debug|Any CPU
+		{11554C34-D81C-4F07-A38D-1E1995014AD2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{11554C34-D81C-4F07-A38D-1E1995014AD2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{11554C34-D81C-4F07-A38D-1E1995014AD2}.Release|x64.ActiveCfg = Release|Any CPU
+		{11554C34-D81C-4F07-A38D-1E1995014AD2}.Release|x64.Build.0 = Release|Any CPU
+		{11554C34-D81C-4F07-A38D-1E1995014AD2}.Release|x86.ActiveCfg = Release|Any CPU
+		{11554C34-D81C-4F07-A38D-1E1995014AD2}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -331,6 +345,7 @@ Global
 		{D32247C4-1645-4DFB-88D2-6BC8D923051A} = {C800B3FC-9890-4858-B2F3-4DA9B6957415}
 		{7F6316D3-3824-4DC6-A68D-01BC82F9B028} = {C800B3FC-9890-4858-B2F3-4DA9B6957415}
 		{D1702287-F937-4429-BF77-4A3862746C1F} = {AAC9998B-F9AF-4096-92CB-81914E85BE56}
+		{11554C34-D81C-4F07-A38D-1E1995014AD2} = {BFF7FB51-DB27-4E46-B575-DDD11430DE8C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {DC32F906-1354-428B-BD36-FC718248DAFB}
